+++ Sources/KeyboardKit/_Deprecated/KeyboardAction+Deprecated.swift
@@ -1,0 +1,7 @@
+import SwiftUI
+
+public extension KeyboardAction {
+
+    @available(*, deprecated, renamed: "KeyboardReturnKeyType")
+    typealias ReturnType = KeyboardReturnKeyType
+}